fix(checkin): Use Next.js cache for event data
Improved performance by caching event data using Next.js's `unstable_cache`.
